+++ packages/api/cms-api/src/redirects/redirects.module.ts
@@ -1,12 +1,13 @@
 import { Block, createOneOfBlock, ExternalLinkBlock, OneOfBlock } from "@comet/blocks-api";
 import { MikroOrmModule } from "@mikro-orm/nestjs";
-import { DynamicModule, Global, Module, ValueProvider } from "@nestjs/common";
+import { DynamicModule, Global, Module, Type, ValueProvider } from "@nestjs/common";
 
 import { InternalLinkBlock } from "../page-tree/blocks/internal-link.block";
 import { RedirectInputFactory } from "./dto/redirect-input.factory";
 import { RedirectEntityFactory } from "./entities/redirect-entity.factory";
 import { createRedirectsResolver } from "./redirects.resolver";
 import { RedirectsService } from "./redirects.service";
+import { RedirectScopeInterface } from "./types";
 
 type CustomTargets = Record<string, Block>;
 
@@ -16,18 +17,19 @@ export const REDIRECTS_LINK_BLOCK = "REDIRECTS_LINK_BLOCK";
 
 interface Config {
     customTargets?: CustomTargets;
+    Scope?: Type<RedirectScopeInterface>;
 }
 @Global()
 @Module({})
 export class RedirectsModule {
-    static register({ customTargets }: Config = {}): DynamicModule {
+    static register({ customTargets, Scope }: Config = {}): DynamicModule {
         const linkBlock = createOneOfBlock(
             { supportedBlocks: { internal: InternalLinkBlock, external: ExternalLinkBlock, ...customTargets }, allowEmpty: false },
             "RedirectsLink",
         );
 
-        const Redirect = RedirectEntityFactory.create(linkBlock);
-        const RedirectInput = RedirectInputFactory.create(linkBlock);
+        const Redirect = RedirectEntityFactory.create({ linkBlock, Scope });
+        const RedirectInput = RedirectInputFactory.create({ linkBlock });
 
         const linkBlockProvider: ValueProvider<RedirectsLinkBlock> = {
             provide: REDIRECTS_LINK_BLOCK,
@@ -37,7 +39,7 @@ export class RedirectsModule {
         return {
             module: RedirectsModule,
             imports: [MikroOrmModule.forFeature([Redirect])],
-            providers: [createRedirectsResolver(Redirect, RedirectInput), RedirectsService, linkBlockProvider],
+            providers: [createRedirectsResolver({ Redirect, RedirectInput, Scope }), RedirectsService, linkBlockProvider],
             exports: [RedirectsService],
         };
     }