+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file. This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [NEXT]
+
+### Highlights
+
+-   Add scope support for redirects (`@comet/cms-api` and `@comet/cms-admin`)
+
 ## 3.0.0
 
 _Oct 17, 2022_